+++ Runtime/UnityReactive/ValueProcessor/MinIntProcessor.cs
@@ -6,11 +6,14 @@ namespace Jeomseon.UnityReactive
     public class MinIntProcessor : IValueProcessor
     {
         [field: SerializeField] public int Min { get; private set; }
+
+        public MinIntProcessor() { }
+
+        public MinIntProcessor(int min) => Min = min;
+
         public T Process<T>(T value)
         {
             return value is int intValue && Mathf.Max(intValue, Min) is T newValue ? newValue : value;
         }
-
-        public MinIntProcessor(int min) => Min = min;
     }
 }