+++ Samples~/BasicUsage/ReactiveListSample.cs
@@ -5,13 +5,13 @@ namespace Jeomseon.Samples.Reactive
 {
     public sealed class ReactiveListSample : MonoBehaviour
     {
-        private readonly ReactiveList<string> _items = new();
+        [SerializeField] private ReactiveList<string> items = new();
 
         private void Awake()
         {
-            _items.AddListenerToAddedEventWithoutNotify(
-                (indices, items) => Debug.Log($"항목 추가: {items[0]}"));
-            _items.Add("Reactive Item");
+            items.AddListenerToAddedEventWithoutNotify(
+                (indices, addedItems) => Debug.Log($"항목 추가: {addedItems[0]}"));
+            items.Add("Reactive Item");
         }
     }
 }